make graph construction its own crate again
