+++ modality-probe-graph/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "modality-probe-graph"
+version = "0.1.0"
+authors = ["dan pittman <dan@auxon.io>"]
+edition = "2018"
+
+[dependencies]
+err-derive = "0.2.4"
+
+modality-probe = { path = "../" }
+modality-probe-collector-common = { path = "../collectors/modality-probe-collector-common" }
+
+[dev-dependencies]
+chrono = "0.4"
